p29.18 chore: internal/ids foundation + reports typed-id alias reconcile
Typed-ID safety capstone, commit 1 of N (owner-approved "everywhere/full").

New leaf package internal/ids defines a DISTINCT Go type for every entity id
(AccountID, SplitID, TransactionID, FundID, ProgramID, SubsidiaryID, UserID,
ReconciliationID, BudgetPlanID, BudgetSplitID, ExpenseReportID,
ExpenseReportLineID, ChangeID, ImportBatchID, ImportRowID, MappingProfileID),
plus generic nullable-FK helpers Ptr/Null for the sql.NullInt64 boundary. It
imports only database/sql, so store/sqlc/reports/web can all import it with no
cycle. A defined int type formats/marshals/compares/scans exactly like int64,
so this is purely compile-time safety: no behavior change.

Reconciles the reports id aliases (compute.go): AccountID/FundID/ProgramID/
SubsidiaryID were `= int64` (zero safety) and are now `= ids.XID`, so a
reports.AccountID IS the distinct type. Typed the four report-owned Params id
fields (Scope/Account/Fund/Program) + ProgramScope + InProgramScope to match;
Reconciliation/Budget stay int64 (their types land in later per-type commits).

The alias flip surfaced the expected reports int64/typed mismatches (more than
"a few" because the toolkit threads store int64 results into map[AccountID]
keys). Fixed by converting int64<->typed ONLY at the two boundaries the
convention names: the store-read seam inside reports (wrap store result fields,
e.g. AccountID(r.AccountID)) and the HTTP-parse seam in web/reports.go
(ids.AccountID(parseID(s))). Drill / store.DrillFilter / indexTree / fundNames
and other int64 helpers stay int64 for now (their per-type commits remove the
scaffolding conversions). Reports tests updated to construct typed Params /
index typed toolkit maps.

Added AGENTS.md Hard Rule 15 documenting the convention (distinct types,
convert only at SQL-scan + HTTP-parse boundaries, non-id int64 stays int64,
nullable FK = sql.NullInt64 at scan / *ids.XID at store via Ptr/Null, audit-row
PKs stay int64, *_versions.EntityID is the typed entity id).

Deviations/notes:
- No PLAN.md step exists for this capstone (it postdates the plan), so no
  checkbox is ticked; recorded here per AGENTS working method.
- PayeeID from the task brief has no backing column (no payees/payee_id table;
  ImportRow.ParsedPayee is a string), so it is intentionally NOT defined.
- sqlc models/store/web bare-int64 ids are unchanged this commit; per-type
  threading (NARROW->PERVASIVE, AccountID last) follows in p29.19+.

make lint (0 issues) + make test (Go + 148 node) + make check all green;
report goldens byte-identical (no testdata churn).

+++ internal/reports/account_ledger.go
@@ -117,8 +117,8 @@ func runAccountLedger(ctx context.Context, tk *Toolkit, p Params) (Table, error)
 		// transactions that produced it (as-of, this account+currency, this scope).
 		open := openByCcy[ccy]
 		openDrill := &Drill{
-			Scope:      p.Scope,
-			AccountIDs: []int64{p.Account},
+			Scope:      int64(p.Scope),
+			AccountIDs: []int64{int64(p.Account)},
 			Currency:   ccy,
 			Mode:       DrillAsOf,
 			AsOf:       opening,
@@ -138,8 +138,8 @@ func runAccountLedger(ctx context.Context, tk *Toolkit, p Params) (Table, error)
 		// scope (descendant closure), ordered (date, split_id) — the SAME filter the
 		// opening/closing balances close over, so opening + Σ == closing.
 		lines, err := tk.Store().DrillSplits(ctx, store.DrillFilter{
-			Scope:     p.Scope,
-			AccountID: p.Account,
+			Scope:     int64(p.Scope),
+			AccountID: int64(p.Account),
 			Currency:  ccy,
 			From:      p.From,
 			To:        p.To,
@@ -166,8 +166,8 @@ func runAccountLedger(ctx context.Context, tk *Toolkit, p Params) (Table, error)
 		// Closing row: the cumulative balance as of To. It EQUALS opening + Σlines
 		// (running) by construction; it is drillable to its own as-of transactions.
 		closeDrill := &Drill{
-			Scope:      p.Scope,
-			AccountIDs: []int64{p.Account},
+			Scope:      int64(p.Scope),
+			AccountIDs: []int64{int64(p.Account)},
 			Currency:   ccy,
 			Mode:       DrillAsOf,
 			AsOf:       p.To,
